Restructure GUI scripts
